Break out websockets
When running clang build analyzer it shows that one of the largest
templates in the Redfish compile unit (which is also the slowest compile
unit) is actually the beast websocket instantiations, taking about 6
seconds to compile on their own.

Luckily, the websocket layer is already split into Websocket and
WebsocketImpl classes that we inherited from crow.  Unfortunately, crow
puts these two classes into the same file.  So.

1. Move the WebSocketImpl class into its own header.
2. Move the websocket() upgrade routes in the websocket routing rule
   into their own compile unit to take advantage of only needing
   WebsocketImpl in this compile unit.

Tested: Drops build time by several seconds depending on what other
level of optimizations are present (1:15 -> 1:00)

[1] https://github.com/aras-p/ClangBuildAnalyzer
